+++ requirements.txt
@@ -1,2 +1,1 @@
 flask>=0.10.1
-git+https://github.com/fzliu/orion-server.git
